Add PHP 8 / PHPUnit 9.0 support
PHPUnit doesn't support PHP 8 until version 9.0, so we need to be
compatible with that while still maintaining compatibility with PHPUnit
5.7.

Fortunately the only thing WorDBLess\BaseTestCase does that's not
compatible with both 5.7 and 9.0 is its use of `setUp` and `tearDown`,
and we can work around that by using `@before` and `@after` instead of
overriding those methods.

Also this adds some tests to verify that the `@before` and `@after`
methods are still called even if a subclass overrides `setUp`/`tearDown`
or supplies its own `@before`/`@after`.