BaseTestCase: Add attributes for PHPUnit 11+ (#75)
PHPUnit 11 has deprecated annotations such as `@before` and `@after` in
favor of attributes like `#[Before]` and `#[After]`. PHPUnit 12 removes
support for the annotations entirely.

We can safely add the attributes even for earlier versions, as PHP 7
sees them as comments and PHP 8 doesn't require that the corresponding
class actually exists.