Fix `get_metadata( $single = true )` when the default is an array
PR #77 attempted to fix `get_metadata( $single = true )` to return the
default metadata rather than always returning an empty string.

Unfortunately the code calling the filter will attempt to unwrap the
return value if the return value is an array; it'll blow up if `$ret[0]`
is not set, or return that instead of the whole default if it does. To
fix it, we can wrap the value if it's an array.

This does re-break `metadata_exists()` for the specific case where the
default metadata is an empty array: normally `metadata_exists()` would
consider that as non-existing, but for WorDBless it will count it as
existing instead. Other than inspecting the call stack, though, I don't
see a way around this.